Move NoticeForm factory to forms.py in pootle_notifications app
